tests/wpt/web-platform-tests/html/browsers/history/the-history-interface/history_pushstate.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>history_pushState</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
      window.history.pushState(1, document.title, '?x=1');
      var state;
      state = window.history.state;
      assert_equals(state, 1, "history state");
    }, "history pushState");
    </script>
  </body>
</html>
